🤖 Move WRITE_DENIED_PREFIX to shared types location
Instead of duplicating the constant between backend (services/tools/fileCommon.ts)
and frontend (constants/ui.ts), moved it to src/types/tools.ts which is already
used as a shared location for tool-related types and constants.

Updated all imports across:
- Backend: file_edit_operation.ts, file_edit_insert.ts, file_edit_operation.test.ts
- Frontend: FileEditToolCall.tsx
- Removed duplicate from constants/ui.ts
- Left comment in fileCommon.ts noting the move

Also updated AGENTS.md with one sentence to prevent this class of mistake:
when a constant exists in a layer-specific location but is needed across
layers, move it to a shared location rather than duplicating it.
